src/toy/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
absl-py==1.2.0
blis==0.4.1
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cymem==2.0.6
google-auth==2.11.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.48.1
gym==0.26.0
gym-notices==0.0.8
idna==3.3
importlib-metadata==4.12.0
joblib==1.2.0
Markdown==3.4.1
MarkupSafe==2.1.1
murmurhash==1.0.8
nltk==3.7
numpy==1.23.3
oauthlib==3.2.1
Pillow==9.2.0
plac==0.9.6
preshed==3.0.7
protobuf==3.19.4
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
spacy==2.2.0
srsly==1.0.5
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==7.1.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
zipp==3.8.1
